Update fs.go
server/sys/fs.go:18:10: invalid operation: stat.Bavail * uint64(stat.Bsize) (mismatched types int64 and uint64)
*** Error code 1

+++ server/sys/fs.go
@@ -15,7 +15,7 @@ import (
 func FreeSpace() (uint64, error) {
 	var stat unix.Statfs_t
 	unix.Statfs(config.Instance().DownloadPath, &stat)
-	return (stat.Bavail * uint64(stat.Bsize)), nil
+	return (uint64(stat.Bavail) * uint64(stat.Bsize)), nil
 }
 
 // Build a directory tree started from the specified path using DFS.
